# swagger-api.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "FinalProject"
  },
  "host": "FinalProject.ru",
  "basePath": "/api/v1",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/store/{store}/order/{order}": {
      "patch": {
        "tags": [
          "common"
        ],
        "summary": "Get all orders in store",
        "description": "description here.",
        "operationId": "updateStatusOrder",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "store",
            "in": "path",
            "description": "Store ID",
            "required": true,
            "type": "integer",
            "format": "int"
          },
          {
            "name": "order",
            "in": "path",
            "description": "Order ID",
            "required": true,
            "type": "integer",
            "format": "int"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}